[autocomment] add setup for cli
