+++ package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "scripts": {
     "build": "tsup src/index.ts --format cjs,esm",
-    "dev": "tsx src/run.ts",
+    "dev": "tsx src/dev.ts",
     "lint": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
     "prettier:check": "prettier . --check",
