+++ src/dev.ts
@@ -1,0 +1,4 @@
+import { main } from './index.js';
+
+// For running locally
+main();
